+++ lib/core/router/app_router.dart
@@ -1,11 +1,11 @@
-import 'package:dutytable/features/calendar/presentation/views/add_calendar_screen.dart';
-import 'package:dutytable/features/calendar/presentation/views/calendar_edit_screen.dart';
+import 'package:dutytable/features/calendar/presentation/views/add/calendar_add_screen.dart';
+import 'package:dutytable/features/calendar/presentation/views/edit/calendar_edit_screen.dart';
 import 'package:dutytable/features/calendar/presentation/views/personal/personal_calendar_screen.dart';
 import 'package:dutytable/features/calendar/presentation/views/shared/shared_calendar_screen.dart';
 import 'package:dutytable/features/notification/presentation/views/notification_screen.dart';
 import 'package:dutytable/features/profile/presentation/views/profile_screen.dart';
 import 'package:dutytable/features/schedule/presentation/views/schedule_add_screen.dart';
-import 'package:dutytable/features/schedule/presentation/views/schedule_detail_screen.dart';
+import 'package:dutytable/features/schedule/presentation/views/detail/schedule_detail_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
@@ -14,8 +14,8 @@ import '../../features/auth/presentation/views/login_screen.dart';
 import '../../features/auth/presentation/views/signup_screen.dart';
 import '../../features/auth/presentation/views/splash_screen.dart';
 import '../../features/calendar/data/models/calendar_model.dart';
-import '../../features/calendar/presentation/views/calendar_setting_screen.dart';
-import '../../features/calendar/presentation/views/shared/shared_calendar_list_screen.dart';
+import '../../features/calendar/presentation/views/setting/calendar_setting_screen.dart';
+import '../../features/calendar/presentation/views/shared/ list/shared_calendar_list_screen.dart';
 import 'app_shell.dart';
 
 GoRouter createRouter(BuildContext context) {
@@ -30,20 +30,13 @@ GoRouter createRouter(BuildContext context) {
       // 회원가입
       GoRoute(path: '/signup', builder: (_, __) => const SignupScreen()),
 
-      // 공유 캘린더 - 캘린더 추가
+      // 공유(개인)캘린더 - 캘린더 추가
       GoRoute(
         path: '/calendar/add',
-        builder: (_, __) => const AddCalendarScreen(),
+        builder: (_, __) => const CalendarAddScreen(),
       ),
 
-      GoRoute(
-        path: '/calendar/setting',
-        builder: (context, state) {
-          final CalendarModel? calendar = state.extra as CalendarModel?;
-          return CalendarSettingScreen(calendar: calendar);
-        },
-      ),
-
+      // 공유(개인)캘린더 - 캘린더 수정
       GoRoute(
         path: '/calendar/edit',
         builder: (context, state) {
@@ -52,19 +45,28 @@ GoRouter createRouter(BuildContext context) {
         },
       ),
 
-      // 알림
+      // 공유(개인)캘린더 - 캘린더 설정
+      GoRoute(
+        path: '/calendar/setting',
+        builder: (context, state) {
+          final CalendarModel? calendar = state.extra as CalendarModel?;
+          return CalendarSettingScreen(calendar: calendar);
+        },
+      ),
+
+      // 공유(개인)캘린더 - 알림
       GoRoute(
         path: '/notification',
         builder: (_, __) => const NotificationScreen(),
       ),
 
-      // 캘린더 - 일정 - 추가
+      // 공유(개인)캘린더 - 일정 - 추가
       GoRoute(
         path: "/schedule/add",
         builder: (_, __) => const ScheduleAddScreen(),
       ),
 
-      // 캘린더 - 일정 - 상세
+      // 공유(개인)캘린더 - 일정 - 상세
       GoRoute(
         path: "/schedule/detail",
         builder: (context, state) {
