+++ lib/features/calendar/presentation/widgets/member_select_dialog/invite_user_tag.dart
@@ -1,0 +1,26 @@
+import 'package:dutytable/core/configs/app_colors.dart';
+import 'package:dutytable/features/calendar/presentation/viewmodels/calendar_add_view_model.dart';
+import 'package:flutter/material.dart';
+
+class InviteUserTag extends StatelessWidget {
+  final CalendarAddViewModel viewModel;
+
+  const InviteUserTag({super.key, required this.viewModel});
+
+  @override
+  Widget build(BuildContext context) {
+    return Wrap(
+      spacing: 8,
+      runSpacing: 4,
+      children: viewModel.invitedUsers.entries
+          .map(
+            (e) => Chip(
+              backgroundColor: AppColors.background(context),
+              label: Text(e.value),
+              onDeleted: () => viewModel.removeInvitedUser(e.key),
+            ),
+          )
+          .toList(),
+    );
+  }
+}
