+++ lib/features/schedule/data/models/address_search_result_model.dart
