+++ lib/features/schedule/data/models/geo_point_model.dart
