+++ lib/features/schedule/data/models/schedule_model.dart
