+++ lib/features/schedule/presentation/viewmodels/schedule_view_model.dart
@@ -1,6 +1,6 @@
 import 'package:dutytable/extensions.dart';
-import 'package:dutytable/features/schedule/datasources/schedule_data_source.dart';
-import 'package:dutytable/features/schedule/models/schedule_model.dart';
+import 'package:dutytable/features/schedule/data/datasources/schedule_data_source.dart';
+import 'package:dutytable/features/schedule/data/models/schedule_model.dart';
 import 'package:dutytable/supabase_manager.dart';
 import 'package:flutter/material.dart';
